# Conflicts:
#	app/src/main/java/com/incon/service/login/LoginDataManagerImpl.java

+++ app/src/main/java/com/incon/service/apimodel/components/login/ServiceCenterResponse.java
@@ -41,7 +41,7 @@ public class ServiceCenterResponse {
     private Integer createdBy;
     @SerializedName("createdDate")
     @Expose
-    private Integer createdDate;
+    private Long createdDate;
     @SerializedName("gstn")
     @Expose
     private String gstn;
@@ -132,11 +132,11 @@ public class ServiceCenterResponse {
         this.createdBy = createdBy;
     }
 
-    public Integer getCreatedDate() {
+    public Long getCreatedDate() {
         return createdDate;
     }
 
-    public void setCreatedDate(Integer createdDate) {
+    public void setCreatedDate(Long createdDate) {
         this.createdDate = createdDate;
     }
